feat: 대회 생성자 및 참여자를 컴포넌트화
