+++ frontend/src/components/CompetitionDetail/AfterCompetition.tsx
@@ -1,8 +1,7 @@
-import { css } from '@style/css';
-
 import { CompetitionInfo } from '@/apis/competitions';
 
 import CompetitionDetailInfo from './CompetitionDetailInfo';
+import CompetitionMembersInfo from './CompetitionMembersInfo';
 import ProblemList from './ProblemList';
 
 interface Props {
@@ -18,8 +17,6 @@ export default function AfterCompetition({
   competition,
   competitionSchedule,
 }: Props) {
-  const formattedParticipants = competition.participants.join(', ');
-
   return (
     <div>
       <CompetitionDetailInfo
@@ -29,34 +26,7 @@ export default function AfterCompetition({
       />
 
       <ProblemList competitionId={competitionId} />
-
-      <div className={containerStyle}>
-        <div className={headerStyle}>대회 생성자</div>
-        <div className={contentStyle}>{competition.host}</div>
-      </div>
-
-      <div className={containerStyle}>
-        <div className={headerStyle}>대회 참여자</div>
-        <div className={contentStyle}>{formattedParticipants}</div>
-      </div>
+      <CompetitionMembersInfo host={competition.host} members={competition.participants} />
     </div>
   );
 }
-
-const containerStyle = css({
-  display: 'flex',
-  flexDirection: 'column',
-  marginTop: '16px',
-  padding: '16px',
-  border: '1px solid #ccc',
-});
-
-const headerStyle = css({
-  fontSize: '18px',
-  fontWeight: 'bold',
-  marginBottom: '8px',
-});
-
-const contentStyle = css({
-  fontSize: '16px',
-});
