+++ frontend/src/components/CompetitionDetail/BeforeCompetition.tsx
@@ -5,6 +5,7 @@ import { CompetitionInfo } from '@/apis/competitions';
 import JoinCompetitionButton from '../Main/Buttons/JoinCompetitionButton';
 import EnterCompetitionButton from './Buttons/EnterCompetitionButton';
 import CompetitionDetailInfo from './CompetitionDetailInfo';
+import CompetitionMembersInfo from './CompetitionMembersInfo';
 
 interface Props {
   competitionId: number;
@@ -22,7 +23,6 @@ export default function BeforeCompetition({
   competitionSchedule,
 }: Props) {
   const BEFORE_COMPETITION_TEXT = ` 시작 전`;
-  const formattedParticipants = competition.participants.join(', ');
 
   return (
     <div>
@@ -37,36 +37,10 @@ export default function BeforeCompetition({
         <EnterCompetitionButton id={competitionId} startsAt={startsAt} endsAt={endsAt} />
       </div>
 
-      <div className={containerStyle}>
-        <div className={headerStyle}>대회 생성자</div>
-        <div className={contentStyle}>{competition.host}</div>
-      </div>
-
-      <div className={containerStyle}>
-        <div className={headerStyle}>대회 참여자</div>
-        <div className={contentStyle}>{formattedParticipants}</div>
-      </div>
+      <CompetitionMembersInfo host={competition.host} members={competition.participants} />
     </div>
   );
 }
-
-const containerStyle = css({
-  display: 'flex',
-  flexDirection: 'column',
-  marginTop: '16px',
-  padding: '16px',
-  border: '1px solid #ccc',
-});
-
-const headerStyle = css({
-  fontSize: '18px',
-  fontWeight: 'bold',
-  marginBottom: '8px',
-});
-
-const contentStyle = css({
-  fontSize: '16px',
-});
 
 const buttonContainerStyle = css({
   display: 'flex',
