+++ frontend/src/components/CompetitionDetail/DuringCompetition.tsx
@@ -1,9 +1,8 @@
-import { css } from '@style/css';
-
 import { CompetitionInfo } from '@/apis/competitions';
 
 import EnterCompetitionButton from './Buttons/EnterCompetitionButton';
 import CompetitionDetailInfo from './CompetitionDetailInfo';
+import CompetitionMembersInfo from './CompetitionMembersInfo';
 
 interface Props {
   competitionId: number;
@@ -22,8 +21,6 @@ export default function DuringCompetition({
   endsAt,
   competitionSchedule,
 }: Props) {
-  const formattedParticipants = competition.participants.join(', ');
-
   return (
     <div>
       <CompetitionDetailInfo
@@ -34,33 +31,7 @@ export default function DuringCompetition({
 
       <EnterCompetitionButton id={competitionId} startsAt={startsAt} endsAt={endsAt} />
 
-      <div className={containerStyle}>
-        <div className={headerStyle}>대회 생성자</div>
-        <div className={contentStyle}>{competition.host}</div>
-      </div>
-
-      <div className={containerStyle}>
-        <div className={headerStyle}>대회 참여자</div>
-        <div className={contentStyle}>{formattedParticipants}</div>
-      </div>
+      <CompetitionMembersInfo host={competition.host} members={competition.participants} />
     </div>
   );
 }
-
-const containerStyle = css({
-  display: 'flex',
-  flexDirection: 'column',
-  marginTop: '16px',
-  padding: '16px',
-  border: '1px solid #ccc',
-});
-
-const headerStyle = css({
-  fontSize: '18px',
-  fontWeight: 'bold',
-  marginBottom: '8px',
-});
-
-const contentStyle = css({
-  fontSize: '16px',
-});
